+++ front-end/JS files/message.js
@@ -1,62 +1,69 @@
-const tbodyTablaMessage = document.getElementById("table_message-tbody");
-const btnRegisterMessage = document.getElementById("btnRegisterMessage");
-const inputMessageId = document.getElementById("message_id");
+const btnCreate = document.getElementById("btnCreate");
+const btnEdit = document.getElementById("btnEdit");
+const btnDelete = document.getElementById("btnDelete");
+const tableBody = document.getElementById("table_body");
 const inputMessageText = document.getElementById("message_messageText");
+const inputMessageOrtopedic = document.getElementById("message_ortopedic");
 
-const urlMessageRequest =
-   "https://g400927313eea0e-g7e55587xh9qd4mr.adb.us-ashburn-1.oraclecloudapps.com/ords/admin/message/message";
+const urlMessageRequest = "http://localhost:8080/api";
 
-// CRUD Message ------------------------------------------------------
+// CRUD MESSAGE ------------------------------------------------------
 
 async function readMessage() {
-   const response = await fetch(urlMessageRequest);
-   const data = await response.json();
-
-   llenarTablaMessage(data.items);
-}
-
-async function createMessage() {
-   console.log("Mensaje: ---");
-
-   console.log(inputMessageText.value);
-
-   const dataToSend = {
-      id: parseInt(inputMessageId.value),
-      messagetext: inputMessageText.value,
-   };
-
    try {
-      const response = await fetch(urlMessageRequest, {
-         method: "POST",
-         body: JSON.stringify(dataToSend),
-         headers: { "Content-type": "application/json" },
-      });
+      const response = await fetch(`${urlMessageRequest}/Message/all`);
+      const data = await response.json();
+      console.log(data);
 
-      if (!response.ok) {
-         throw new Error("Error with POST...");
-      }
+      // sessionStorage.setItem("dataClient", JSON.stringify(data));
 
-      location.reload();
+      llenarTabla(data);
    } catch (err) {
       console.log(err);
    }
 }
 
-async function updateMessage(filaEditada) {
+async function createMessage() {
+   if (inputMessageText.value !== "") {
+      const dataToSend = {
+         messageText: inputMessageText.value,
+      };
+
+      try {
+         const response = await fetch(`${urlMessageRequest}/Message/save`, {
+            method: "POST",
+            body: JSON.stringify(dataToSend),
+            headers: { "Content-type": "application/json" },
+         });
+
+         if (!response.ok) {
+            throw new Error("Error with post");
+         }
+
+         location.reload();
+      } catch (err) {
+         console.log(err);
+      }
+   }
+}
+
+async function updateMessage(filaEditada, id) {
    const dataToSend = {
-      id: parseInt(filaEditada[0].innerText),
-      messagetext: filaEditada[1].innerText,
+      idMessage: id,
+      messageText: filaEditada[0].innerText,
    };
 
+   console.log(dataToSend);
+
    try {
-      const response = await fetch(urlMessageRequest, {
+      const response = await fetch(`${urlMessageRequest}/Message/update`, {
          method: "PUT",
          body: JSON.stringify(dataToSend),
          headers: { "Content-type": "application/json" },
       });
 
       if (!response.ok) {
-         throw new Error("Error with put...");
+         throw new Error("Error with put...:" + response.statusText);
       }
 
       location.reload();
@@ -65,9 +72,21 @@ async function updateMessage(filaEditada) {
    }
 }
 
-async function deleteMessage(id) {
+async function deleteMessage() {
+   const rowSelected = Array.from(tableBody.children).find((row) => {
+      return row.dataset.rowIsSelected === "true";
+   });
+
+   if (!rowSelected) {
+      alert("Please select a message to delete");
+      return;
+   }
+
+   const id = rowSelected.dataset.id;
+   console.log(id);
+
    try {
-      const response = await fetch(`${urlMessageRequest}/${id}`, {
+      const response = await fetch(`${urlMessageRequest}/Message/${id}`, {
          method: "DELETE",
          body: null,
       });
@@ -84,51 +103,91 @@ async function deleteMessage(id) {
 
 readMessage();
 
-const llenarTablaMessage = (data) => {
-   data.forEach((item) => {
-      const newMessage = tbodyTablaMessage.insertRow();
+const llenarTabla = (data) => {
+   if (data) {
+      data.forEach((item) => {
+         const newRow = `
+          <tr class="tr_table" data-id=${item.idMessage} >
+            <td>${item.messageText}</td>
+            <td>-- ortopedics name --</td>
+          </tr>
+         `;
 
-      newMessage.insertCell([0]).innerHTML = item.id;
-      newMessage.insertCell([1]).innerHTML = item.messagetext;
-
-      newMessage.insertCell([
-         2,
-      ]).innerHTML = `<button onClick="habilitarEdicionTablaMessage(this)">Edit</button>`;
-      newMessage.insertCell([
-         3,
-      ]).innerHTML = `<button onClick="deleteMessage(${item.id})">Delete</button> `;
-   });
+         tableBody.insertAdjacentHTML("beforeend", newRow);
+      });
+   }
 };
 
-const habilitarEdicionTablaMessage = (eventTarget) => {
-   const tdContenedorEditBtn = eventTarget.parentElement;
-   const filaAEditar = eventTarget.parentElement.parentElement;
-   const arrayFilaAEditar = Array.from(filaAEditar.children);
+const effectsOnRows = (event) => {
+   const rowSelected = event.target.closest(".tr_table");
 
-   arrayFilaAEditar.forEach((elem, index) => {
-      if (
-         index !== 0 &&
-         index !== arrayFilaAEditar.length - 1 &&
-         index !== arrayFilaAEditar.length - 2
-      )
+   if (event.type === "click") {
+      tableBody.removeEventListener("mouseover", effectsOnRows);
+      tableBody.removeEventListener("mouseout", effectsOnRows);
+
+      Array.from(tableBody.children).forEach((el) => {
+         el.dataset.rowIsSelected = "false";
+         el.style.color = "#000";
+         el.style.outline = "";
+         el.style["outline-offset"] = "";
+      });
+
+      rowSelected.dataset.rowIsSelected = "true";
+      rowSelected.style.color = "rgb(149,52,10)";
+      rowSelected.style.outline = "2px solid rgb(231, 78, 12)";
+      rowSelected.style["outline-offset"] = "-3px";
+
+      return;
+   }
+   if (event.type === "mouseover") {
+      rowSelected.style.outline = "2px solid rgb(231, 78, 12)";
+      rowSelected.style["outline-offset"] = "-3px";
+      return;
+   }
+   if (event.type === "mouseout") {
+      rowSelected.style.outline = "";
+      rowSelected.style["outline-offset"] = "";
+      return;
+   }
+};
+
+const habilitarEdicionTabla = (event) => {
+   const rowSelected = Array.from(tableBody.children).find((row) => {
+      return row.dataset.rowIsSelected === "true";
+   });
+
+   if (!rowSelected) {
+      alert("Please select a message to edit");
+      return;
+   }
+
+   if (event.target.dataset.type === "edit") {
+      Array.from(rowSelected.children).forEach((elem, index) => {
          elem.setAttribute("contenteditable", "");
-      if (index === 1) elem.focus();
-   });
+         if (index === 0) elem.focus();
+      });
 
-   crearBtnGuardarMessage(tdContenedorEditBtn);
+      btnEdit.innerText = "Save";
+      event.target.dataset.type = "save";
+      return;
+   }
+
+   const rowEditedId = parseInt(rowSelected.dataset.id);
+
+   updateMessage(Array.from(rowSelected.children), rowEditedId);
 };
 
-const crearBtnGuardarMessage = (contenedorBtn) => {
-   const btnGuardar = document.createElement("button");
-   btnGuardar.textContent = "Save";
-   contenedorBtn.innerHTML = "";
-   contenedorBtn.prepend(btnGuardar);
-   btnGuardar.addEventListener("click", () => {
-      updateMessage(Array.from(contenedorBtn.parentElement.cells));
-   });
-};
+// Event listeners **************************************************
 
-btnRegisterMessage.addEventListener("click", (event) => {
+btnCreate.addEventListener("click", (event) => {
    event.preventDefault();
    createMessage();
 });
+
+btnEdit.addEventListener("click", habilitarEdicionTabla);
+
+btnDelete.addEventListener("click", deleteMessage);
+
+tableBody.addEventListener("click", effectsOnRows);
+tableBody.addEventListener("mouseover", effectsOnRows);
+tableBody.addEventListener("mouseout", effectsOnRows);
